Convert ssh to use a pty when executing a remote command
Using a pty also means that we need to signal EOF explicitly by sending
a Ctrl-D to the remote process.

What this patch also does it to clear handle->channel at the very end
of executing the command. This also prevented the signal handling from
ever being able to work :-)

Signed-off-by: Olaf Kirch <okir@suse.de>