Preparations for timeout management: ruby wrappers
Also fixed a build bug, but the fix will be incomplete without
a change to the spec files to create a /usr/lib64/libtwopence.so
(with no version) symlink.

Also made some parameters in the ruby API optional. This will
introduce an incompatibility in examples (already fixed), in the
wicked testsuite and the pennyworth bindings
(Christian has been made aware).