Add a workaround for a virtio-serial issue
The virtio-serial port will always return POLLHUP as long as nothing is
connected to the host side socket. Normally, we interpret this as an
indication to shut down the guest side port, but that's not a very
productive thing to do.

For now, we still treat POLLHUP as meaning: "close the device", but we will
reopen it right away and wait for someone else to connect. This is far from
optimal, but it's a band aid for now.

Signed-off-by: Olaf Kirch <okir@suse.de>