Fix "make tests" target
Commit 869fb560 moved files around, and introduced a makefile
in tests/. Instead of executing the tests with the top twopence
source directory as current working directory, the tests are
now executed with PWD=$topdir/tests. This caused LD_LIBRARY_PATH
to be set incorrecty, causing all tests to fail if libtwopence
is not in the installed system.

This patch fixes this issue.

Another change introduced by this patch is to silently skip the
ruby tests if rspec is not installed.

Signed-off-by: Olaf Kirch <okir@suse.de>