client/server: fix extraction of /proc files
When you stat a /proc file, it will always report a size of 0.
This doesn't work with the way extraction used to work, where the
server informs the client of the file size before transmitting it.

This patch changes the protocol in the following way:

Client sends the server an 'e' packet as before, specifying the
file and user name.

Server responds with one of the following:

 - in case of an error: 'M' (major status) packet containing the error code
 - otherwise, a stream of 'd' (data) packets containing the file data.
   The end of the file transfer is indicated by an 'E' (EOF) packet

Signed-off-by: Olaf Kirch <okir@suse.de>