server: fix a timing related protocol bug
Due to bad timing, we may receive the stdin EOF indication from the client
after the process as exited. In this case, the transaction may no longer exist,
which would cause us to send an EPROTO status.

The client would thus see a duplicate status packet, causing it to bail out
with an error even though the transaction succeeded.

Signed-off-by: Olaf Kirch <okir@suse.de>