Another attempt at working around the deficiencies of ssh signal propagation
What we do now is
 - catch signal
 - mark foreground transaction as interrupted
 - in the event loop, detect that and shut down the
   session associated with this transaction

Unfortunately, session shutdown seems to be the only option right now.
This is sad, because it means we cannot share sessions across commands.
I also considered just closing the channel with ssh_channel_close(),
but that will block waiting for the remote command to finish.

Signed-off-by: Olaf Kirch <okir@suse.de>