socket: fix a poll/doio issue with ttys
When we receive a POLLHUP, we always also try to read from the descriptor.
This is needed in order to pull in residual data that may be stuck in a
kernel buffer.

However, this becomes problematic with ptys - the slave pty will
return EIO when trying to read it even though it's been hung up.
Work around this by ignoring read errors when POLLHUP is set.

Signed-off-by: Olaf Kirch <okir@suse.de>