Introduce new twopence_disconnect(target) function
Quoting from the manpage:

Disconnecting from the System Under Test

When using twopence_target_free(3) to destroy the target handle,
all state on pending transactions, etc, will also be lost. A less
destructive approach is to disconnect from the SUT but otherwise
retain the target handle:

  int twopence_disconnect(twopence_target_t *target);

This will drop the underlying transport connection, and cancel all
pending transactions.
It will be possible to reap the status of pending commands
after this, but all attempts to interact with the remote
system will return TWOPENCE_TRANSPORT_ERROR.

Signed-off-by: Olaf Kirch <okir@suse.de>